pyproject.toml
[project]
name = "fastapi-starter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic >= 1.18.3",
    "anyio >= 4.8.0",
    "asyncpg >= 0.31.0",
    "click>=8.3.1",
    "fastapi-mail>=1.6.1",
    "fastapi-users[sqlalchemy]>=15.0.4",
    "fastapi[standard-no-fastapi-cloud-cli] >= 0.128.4",
    "httpx >= 0.28.1",
    "orjson>=3.11.7",
    "pydantic>=2.10.6",
    "resend>=2.21.0",
    "sqladmin[full]>=0.23.0",
    "sqlalchemy[asyncio] >= 2.0.46",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "pyqt6>=6.10.2",
    "pytest >= 9.0.2",
    "pytest-xdist>=3.8.0",
    "python-dotenv>=1.2.1",
    "qtconsole>=5.7.1",
    "ruff >= 0.9.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["cli.py"]

[project.scripts]
cli = "cli:main"

[tool.ruff]
line-length = 90
target-version = "py312"

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint.isort]
case-sensitive = true
lines-after-imports = 2
lines-between-types = 1
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
    "tests",
]

[tool.ruff.lint.isort.sections]
tests = ["tests"]

[tool.pytest.ini_options]
filterwarnings = [

]
